+++ com.unity.render-pipelines.high-definition/CHANGELOG.md
@@ -154,6 +154,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added scene view exposure override.
 - Added support for exposure curve remapping for min/max limits.
 - Added presets for ray traced reflections.
+- Added final image histogram debug view (both luminance and RGB).
 
 ### Fixed
 - Fix when rescale probe all direction below zero (1219246)